requirements.txt
httpx>=0.27
python-dotenv>=1.0
feedparser>=6.0
google-genai>=1.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
